% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.apply_group_filter}
\alias{.apply_group_filter}
\title{Apply group-level missing-rate filter}
\usage{
.apply_group_filter(merger, df)
}
\description{
Apply group-level missing-rate filter
}
\keyword{internal}
