% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.basic_cols}
\alias{.basic_cols}
\title{Return the subset of site-metadata column names present in df}
\usage{
.basic_cols(df)
}
\description{
Return the subset of site-metadata column names present in df
}
\keyword{internal}
