% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.calc_mod_rate}
\alias{.calc_mod_rate}
\title{Compute per-site modification rate from pileup base counts}
\usage{
.calc_mod_rate(df, method, custom_func = NULL)
}
\description{
Compute per-site modification rate from pileup base counts
}
\keyword{internal}
