% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_core.R
\name{.calculate_metagene_series}
\alias{.calculate_metagene_series}
\title{Calculate one metagene profile series}
\usage{
.calculate_metagene_series(
  dt,
  bin_number,
  value_col = NULL,
  smooth = TRUE,
  span = 0.3
)
}
\description{
Calculate one metagene profile series
}
\keyword{internal}
