% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_cols}
\alias{.check_cols}
\title{Assert that required columns exist in a data frame}
\usage{
.check_cols(df, required, df_name = "df")
}
\description{
Assert that required columns exist in a data frame
}
\keyword{internal}
