% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_count}
\alias{.check_count}
\title{Assert that a value is a positive integer scalar}
\usage{
.check_count(x, name)
}
\description{
Assert that a value is a positive integer scalar
}
\keyword{internal}
