% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.check_merger_args}
\alias{.check_merger_args}
\title{Validate new_merger() arguments}
\usage{
.check_merger_args(
  sample_files,
  condition,
  modification_method,
  sample_names,
  min_modification_rate,
  min_depth,
  group_missing_threshold,
  min_group_mean_rate,
  group_filter_strategy,
  custom_calc_func
)
}
\description{
Validate new_merger() arguments
}
\keyword{internal}
