% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.combine_site_metadata}
\alias{.combine_site_metadata}
\title{Combine site metadata across processed sample tables}
\usage{
.combine_site_metadata(sample_tables)
}
\description{
Combine site metadata across processed sample tables
}
\keyword{internal}
