% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_heights.R
\name{.compute_bits}
\alias{.compute_bits}
\title{Compute bits (information content) per position}
\usage{
.compute_bits(pfm, N = 4L, Nseqs = NULL)
}
\description{
Compute bits (information content) per position
}
\keyword{internal}
