% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ensure_dir}
\alias{.ensure_dir}
\title{Create output directory, warning on failure but not stopping}
\usage{
.ensure_dir(path)
}
\description{
Create output directory, warning on failure but not stopping
}
\keyword{internal}
