% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.filter_by_group_mean}
\alias{.filter_by_group_mean}
\title{Keep only sites where at least one group has mean rate >= threshold}
\usage{
.filter_by_group_mean(merger, df)
}
\description{
Keep only sites where at least one group has mean rate >= threshold
}
\keyword{internal}
