% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_heights.R
\name{.find_namespace}
\alias{.find_namespace}
\title{Resolve namespace from inputs}
\usage{
.find_namespace(letter_mat, seq_type, namespace)
}
\description{
Resolve namespace from inputs
}
\keyword{internal}
