% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_col_schemes.R
\name{.get_col_scheme}
\alias{.get_col_scheme}
\title{Retrieve a built-in or user-defined colour scheme}
\usage{
.get_col_scheme(col_scheme, seq_type = "auto")
}
\arguments{
\item{col_scheme}{Name of a built-in scheme (see [list_col_schemes()]) or a
`modsite_cs` object created by [make_col_scheme()].}

\item{seq_type}{Sequence type, used only when `col_scheme = "auto"`.}
}
\value{
A `modsite_cs` `data.frame`.
}
\description{
Retrieve a built-in or user-defined colour scheme
}
\keyword{internal}
