% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_logo.R
\name{.get_font}
\alias{.get_font}
\title{Load a font from the package's extdata directory (with option-level cache)}
\usage{
.get_font(font)
}
\arguments{
\item{font}{Font name.  See [list_fonts()].}
}
\value{
A `data.frame` of polygon glyph data for the requested font.
}
\description{
Load a font from the package's extdata directory (with option-level cache)
}
\keyword{internal}
