% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.group_indices}
\alias{.group_indices}
\title{Determine group indices (list of integer vectors, one per group)}
\usage{
.group_indices(merger)
}
\description{
Determine group indices (list of integer vectors, one per group)
}
\keyword{internal}
