% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_heights.R
\name{.guess_seq_type}
\alias{.guess_seq_type}
\title{Guess sequence type from the letters present}
\usage{
.guess_seq_type(letter_vec)
}
\description{
Guess sequence type from the letters present
}
\keyword{internal}
