% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger_visualization.R
\name{.impute_vector_by_group}
\alias{.impute_vector_by_group}
\title{Impute missing values within groups}
\usage{
.impute_vector_by_group(x, groups, method = c("mean", "median"))
}
\description{
Impute missing values within groups
}
\keyword{internal}
