% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_heights.R
\name{.letter_matrix}
\alias{.letter_matrix}
\title{Build a character matrix from a vector of equal-length sequences}
\usage{
.letter_matrix(seqs)
}
\description{
Build a character matrix from a vector of equal-length sequences
}
\keyword{internal}
