% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.load_sample_file}
\alias{.load_sample_file}
\title{Load a single pileup file into a data frame}
\usage{
.load_sample_file(file_path)
}
\description{
Load a single pileup file into a data frame
}
\keyword{internal}
