% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.metagene_default_title}
\alias{.metagene_default_title}
\title{Default title for metagene plots}
\usage{
.metagene_default_title(cols)
}
\description{
Default title for metagene plots
}
\keyword{internal}
