% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_grouped.R
\name{.metagene_group_numeric}
\alias{.metagene_group_numeric}
\title{Group numeric values using a cutoff or breaks}
\usage{
.metagene_group_numeric(
  x,
  group_by,
  group_cutoff = 0.05,
  group_breaks = NULL,
  group_labels = NULL
)
}
\description{
Group numeric values using a cutoff or breaks
}
\keyword{internal}
