% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_grouped.R
\name{.metagene_group_values}
\alias{.metagene_group_values}
\title{Create grouping labels for grouped metagene analysis}
\usage{
.metagene_group_values(
  x,
  group_by,
  group_cutoff = 0.05,
  group_breaks = NULL,
  group_labels = NULL,
  include_na_group = FALSE
)
}
\description{
Create grouping labels for grouped metagene analysis
}
\keyword{internal}
