% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_grouped.R
\name{.metagene_grouped_default_title}
\alias{.metagene_grouped_default_title}
\title{Default title for grouped metagene plots}
\usage{
.metagene_grouped_default_title(spec)
}
\description{
Default title for grouped metagene plots
}
\keyword{internal}
