% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_grouped.R
\name{.metagene_grouped_y_label}
\alias{.metagene_grouped_y_label}
\title{Y-axis label for grouped metagene plots}
\usage{
.metagene_grouped_y_label(spec)
}
\description{
Y-axis label for grouped metagene plots
}
\keyword{internal}
