% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_grouped.R
\name{.metagene_interval_label}
\alias{.metagene_interval_label}
\title{Create default interval labels}
\usage{
.metagene_interval_label(lo, hi, idx, n_groups)
}
\description{
Create default interval labels
}
\keyword{internal}
