% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_grouped.R
\name{.metagene_minus_log10}
\alias{.metagene_minus_log10}
\title{Compute a stable -log10 transform for p-values}
\usage{
.metagene_minus_log10(x)
}
\description{
Compute a stable -log10 transform for p-values
}
\keyword{internal}
