% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.metagene_profile_columns}
\alias{.metagene_profile_columns}
\title{Return metagene profile columns}
\usage{
.metagene_profile_columns(profile_data)
}
\description{
Return metagene profile columns
}
\keyword{internal}
