% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.metagene_series_labels}
\alias{.metagene_series_labels}
\title{Human-readable labels for metagene series}
\usage{
.metagene_series_labels(cols)
}
\description{
Human-readable labels for metagene series
}
\keyword{internal}
