% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.normalize_metagene_weight_cols}
\alias{.normalize_metagene_weight_cols}
\title{Normalise metagene weight columns}
\usage{
.normalize_metagene_weight_cols(weight_cols, df)
}
\description{
Normalise metagene weight columns
}
\keyword{internal}
