% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_core.R
\name{.prepare_metagene_bins}
\alias{.prepare_metagene_bins}
\title{Prepare binned metagene positions}
\usage{
.prepare_metagene_bins(mapped_res, bin_number = 100L)
}
\description{
Prepare binned metagene positions
}
\keyword{internal}
