% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genome.R
\name{.print_region_summary}
\alias{.print_region_summary}
\title{Print a genomic-region annotation summary to the console}
\usage{
.print_region_summary(df)
}
\description{
Print a genomic-region annotation summary to the console
}
\keyword{internal}
