% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.process_one_sample}
\alias{.process_one_sample}
\title{Process one sample: load, compute mod-rate, apply site-level filters}
\usage{
.process_one_sample(merger, file_path, sample_name)
}
\description{
Process one sample: load, compute mod-rate, apply site-level filters
}
\keyword{internal}
