% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{.remove_all_zero_rows}
\alias{.remove_all_zero_rows}
\title{Remove sites where all sample rates are 0 or NA}
\usage{
.remove_all_zero_rows(merger, df)
}
\description{
Remove sites where all sample rates are 0 or NA
}
\keyword{internal}
