% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.resolve_metagene_sites_df}
\alias{.resolve_metagene_sites_df}
\title{Resolve metagene site input}
\usage{
.resolve_metagene_sites_df(sites_df = NULL, annotated_df = NULL)
}
\description{
Resolve metagene site input
}
\keyword{internal}
