% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_core.R
\name{.smooth_metagene_values}
\alias{.smooth_metagene_values}
\title{Smooth metagene profile values}
\usage{
.smooth_metagene_values(values, bins, smooth = TRUE, span = 0.3)
}
\description{
Smooth metagene profile values
}
\keyword{internal}
