% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{genomic_region_stats}
\alias{genomic_region_stats}
\title{Genomic region statistics}
\usage{
genomic_region_stats(x)
}
\arguments{
\item{x}{A `SampleStats` object created by [new_sample_stats()].}
}
\value{
A `data.frame` with one row per (sample, genomic_region).
}
\description{
Requires an annotation column `genomic_region`.
}
\examples{
df <- data.frame(chrom = "chr1", pos = 1:4, ref = "A",
                 genomic_region = c("CDS","UTR3","CDS","intron"),
                 s1 = c(0.1, 0.2, 0.3, 0.05))
ss <- new_sample_stats(df, sample_cols = "s1")
genomic_region_stats(ss)

}
