% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R, R/metagene_core.R
\name{\%||\%}
\alias{\%||\%}
\title{Provide a default when x is NULL}
\usage{
x \%||\% y

x \%||\% y
}
\description{
Provide a default when x is NULL

Provide a default when x is NULL
}
\keyword{internal}
