% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_col_schemes.R
\name{motif_col_schemes}
\alias{motif_col_schemes}
\title{Sequence logo colour schemes}
\description{
Built-in colour schemes for sequence logos plus utilities for creating
custom schemes.

Adapted from the ggseqlogo package (Wagih 2017).
}
