% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{new_saturation_analyzer}
\alias{new_saturation_analyzer}
\title{Saturation analysis settings}
\usage{
new_saturation_analyzer(
  bam_file,
  min_depth = 10L,
  min_mod_rate = 0.05,
  min_base_qual = 20L,
  test_mode = TRUE,
  max_sites = 100000L
)
}
\arguments{
\item{bam_file}{Path to a BAM file. Must exist when calling
[extract_site_stats()]. For unit tests or advanced usage, you can construct
the object and directly set `analyzer$site_stats`.}

\item{min_depth}{Integer >= 0. Minimum depth to call a site "valid".}

\item{min_mod_rate}{Numeric scalar in [0, 1]. Minimum modification rate to
call a site "valid".}

\item{min_base_qual}{Integer >= 0. Minimum base quality for pileup counting.}

\item{test_mode}{Logical. If `TRUE`, only a limited number of sites are
processed to reduce runtime.}

\item{max_sites}{Integer >= 0. Maximum number of sites to process when
`test_mode = TRUE`.}
}
\value{
An object of class `SaturationAnalyzer`.
}
\description{
Construct a saturation analyzer. The typical workflow is:
[new_saturation_analyzer()] -> [extract_site_stats()] ->
[simulate_downsampling()] -> [plot_saturation()] or
[run_saturation_analysis()].
}
\details{
The BAM parsing requires Bioconductor package `Rsamtools`, which is placed in
`Suggests` and loaded on demand.
}
\examples{
\dontrun{
az <- new_saturation_analyzer("sample.bam", min_depth = 10L,
                              min_mod_rate = 0.05, test_mode = TRUE)
}

}
