% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_logo.R
\name{theme_logo}
\alias{theme_logo}
\title{ggplot2 theme for sequence logos}
\usage{
theme_logo(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.  Default 12.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
A minimal theme with suppressed grid lines and black axis text.
}
